+++ src/main/java/com/zxl/materialStorage/controller/materialEnter/MaterialPackingController.java
@@ -1,15 +1,22 @@
 package com.zxl.materialStorage.controller.materialEnter;
 
 import cn.hutool.core.util.ObjectUtil;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.zxl.materialStorage.common.api.ApiResult;
+import com.zxl.materialStorage.model.pojo.MaterialEnter;
 import com.zxl.materialStorage.model.pojo.MaterialPacking;
+import com.zxl.materialStorage.model.pojo.MaterialType;
+import com.zxl.materialStorage.service.materialEnter.MaterialEnterService;
 import com.zxl.materialStorage.service.materialEnter.MaterialPackingService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @className: MaterialPackingController
@@ -23,6 +30,8 @@ import java.util.List;
 public class MaterialPackingController {
     @Autowired
     private MaterialPackingService materialPackingService;
+    @Autowired
+    private MaterialEnterService materialEnterService;
 
     @PostMapping("/insertNewOne")
     public ApiResult<Object> insertNewOne(@RequestBody MaterialPacking materialPacking){
@@ -38,12 +47,17 @@ public class MaterialPackingController {
         return ApiResult.success();
     }
 
-    //TODO 若存在引用则拒绝删除
     @DeleteMapping("/deleteOne")
     public ApiResult<Object> deleteOne(@RequestParam(value = "empId") String empId){
         try {
             if (StringUtils.isEmpty(empId)){
                 return ApiResult.blank();
+            }
+            //存在引用，拒绝删除，直接返回
+            MaterialPacking byId = materialPackingService.getById(empId);
+            MaterialEnter existMaterialEnter = materialEnterService.getOne(new QueryWrapper<MaterialEnter>().lambda().eq(MaterialEnter::getEmpNo, byId.getEmpNo()));
+            if (ObjectUtil.isNotNull(existMaterialEnter)){
+                return ApiResult.error();
             }
             materialPackingService.deleteOne(empId);
         } catch (Exception e) {
@@ -53,12 +67,22 @@ public class MaterialPackingController {
         return ApiResult.success();
     }
 
-    //TODO 若存在引用则拒绝删除
     @DeleteMapping("/deleteMany")
     public ApiResult<Object> deleteMany(@RequestBody List<String> empIdList){
         try {
             if (ObjectUtil.isEmpty(empIdList)){
                 return ApiResult.blank();
+            }
+            List<MaterialPacking> materialPackingList = materialPackingService.listByIds(empIdList);
+            List<String> empNoList = new ArrayList<>();
+            for (MaterialPacking materialPacking : materialPackingList) {
+                empNoList.add(materialPacking.getEmpNo());
+            }
+            Map<String, Object> columnMap = new HashMap<>();
+            columnMap.put("emp_no",empNoList);
+            List<MaterialEnter> materialEnters = materialEnterService.listByMap(columnMap);
+            if (ObjectUtil.isNotEmpty(materialEnters)){
+                return ApiResult.error();
             }
             materialPackingService.deleteMany(empIdList);
         } catch (Exception e) {
@@ -75,6 +99,12 @@ public class MaterialPackingController {
             if (ObjectUtil.isNull(materialPacking)){
                 return ApiResult.blank();
             }
+            //异步更新引用
+            MaterialPacking byId = materialPackingService.getById(materialPacking.getEmpId());
+            if (!byId.getEmpNo().equals(materialPacking.getEmpNo())){
+                materialEnterService.updateEmpNos(byId,materialPacking);
+            }
+            //再更新自己
             materialPackingService.updateOne(materialPacking);
         } catch (Exception e) {
             log.error("更新物资打包方式出错");
